Avoid removing all contents when formatting file with syntax error (#314)
Exploring a resolution to
https://github.com/astral-sh/ruff-vscode/issues/336

It looks like at
https://github.com/astral-sh/ruff-lsp/compare/v0.0.42...v0.0.43#diff-9cc81288343e2240bb3387d88ab41ebf706e40c86af24cf96bad8919a48d5cceL762-L763
we no longer exit early if stdout is empty. This pull request does not
restore that logic, but rather adds a return code check and exits if
there is a bad return code.

<!--
Thank you for contributing to Ruff! To help us out with reviewing,
please consider the following:

- Does this pull request include a summary of the change? (See below.)
- Does this pull request include a descriptive title?
- Does this pull request include references to any relevant issues?
-->

## Summary

- Add failing test case for formatting code with a syntax error
- Track return code of `ruff format`
- Do not format documents with non-zero return code from Ruff

<!-- What's the purpose of the change? What does it do, and why? -->

## Test Plan

<!-- How was it tested? -->