Avoid writing empty source on excluded files (#317)
Right now, if the user tries to fix or format an excluded file, we don't
write anything to `stdout`. As a result, the LSP then overwrites the
file with the empty string.

This is a compromise fix whereby we avoid writing empty fixes in the
LSP. It's a "compromise" in that it'll do the wrong thing in some cases,
e.g., if the file is _just_ `import os`, we won't remove that when the
user runs "Fix all". However, I think we already had this behavior in
the LSP in the past, and we can fix it for newer versions by changing
Ruff to output the unchanged file.

## Test Plan

Ran the debug extension.

Opened a first-party file, and ensured that formatting and linting work
as before (with the exception of the `import os` case).

Opened a third-party file, and ensured that formatting and linting had
no effect (and didn't delete the file).